// generated/models.ts
export type Access = {
    data: AccessData
    delete: boolean
    externalize: boolean
    manage: boolean
    read: boolean
    update: boolean
    write: boolean
}

export type AccessData = {
    read: boolean
    write: boolean
}
// ...
    export enum type {
        PERSONAL_ACCESS_TOKEN_V1 = 'PERSONAL_ACCESS_TOKEN_V1',
        PERSONAL_ACCESS_TOKEN_V2 = 'PERSONAL_ACCESS_TOKEN_V2',
    }
// ...
    export enum valueType {
        TEXT = 'TEXT',
        LONG_TEXT = 'LONG_TEXT',
        MULTI_TEXT = 'MULTI_TEXT',
        LETTER = 'LETTER',
        PHONE_NUMBER = 'PHONE_NUMBER',
        EMAIL = 'EMAIL',
        BOOLEAN = 'BOOLEAN',
        TRUE_ONLY = 'TRUE_ONLY',
        DATE = 'DATE',
        DATETIME = 'DATETIME',
        TIME = 'TIME',
        NUMBER = 'NUMBER',
        UNIT_INTERVAL = 'UNIT_INTERVAL',
        PERCENTAGE = 'PERCENTAGE',
        INTEGER = 'INTEGER',
        INTEGER_POSITIVE = 'INTEGER_POSITIVE',
        INTEGER_NEGATIVE = 'INTEGER_NEGATIVE',
        INTEGER_ZERO_OR_POSITIVE = 'INTEGER_ZERO_OR_POSITIVE',
        TRACKER_ASSOCIATE = 'TRACKER_ASSOCIATE',
        USERNAME = 'USERNAME',
        COORDINATE = 'COORDINATE',
        ORGANISATION_UNIT = 'ORGANISATION_UNIT',
        REFERENCE = 'REFERENCE',
        AGE = 'AGE',
        URL = 'URL',
        FILE_RESOURCE = 'FILE_RESOURCE',
        IMAGE = 'IMAGE',
        GEOJSON = 'GEOJSON',
    }
// ...
export type AttributeValue = {
    /**
     * A UID reference to a Attribute
     * (Java name `org.hisp.dhis.attribute.Attribute`)
     */
    attribute?: {
        id: string
    }
    value: string
}
// ...
    export enum aggregationType {
        SUM = 'SUM',
        AVERAGE = 'AVERAGE',
        AVERAGE_SUM_ORG_UNIT = 'AVERAGE_SUM_ORG_UNIT',
        LAST = 'LAST',
        LAST_AVERAGE_ORG_UNIT = 'LAST_AVERAGE_ORG_UNIT',
        LAST_LAST_ORG_UNIT = 'LAST_LAST_ORG_UNIT',
        LAST_IN_PERIOD = 'LAST_IN_PERIOD',
        LAST_IN_PERIOD_AVERAGE_ORG_UNIT = 'LAST_IN_PERIOD_AVERAGE_ORG_UNIT',
        FIRST = 'FIRST',
        FIRST_AVERAGE_ORG_UNIT = 'FIRST_AVERAGE_ORG_UNIT',
        FIRST_FIRST_ORG_UNIT = 'FIRST_FIRST_ORG_UNIT',
        COUNT = 'COUNT',
        STDDEV = 'STDDEV',
        VARIANCE = 'VARIANCE',
        MIN = 'MIN',
        MAX = 'MAX',
        MIN_SUM_ORG_UNIT = 'MIN_SUM_ORG_UNIT',
        MAX_SUM_ORG_UNIT = 'MAX_SUM_ORG_UNIT',
        NONE = 'NONE',
        CUSTOM = 'CUSTOM',
        DEFAULT = 'DEFAULT',
    }

    export enum dataDimensionType {
        DISAGGREGATION = 'DISAGGREGATION',
        ATTRIBUTE = 'ATTRIBUTE',
    }
// ...
export type CategoryCombo = {
    access: Access
    attributeValues: Array<AttributeValue>
    categories?: Array<{
        id: string
    }>
    categoryOptionCombos?: Array<{
        id: string
    }>
    code: string
    created: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    createdBy?: {
        id: string
    }
    dataDimensionType: CategoryCombo.dataDimensionType
    displayName: string
    favorite: boolean
    favorites: Array<string>
    href: string
    id: string
    isDefault: boolean
    lastUpdated: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    lastUpdatedBy?: {
        id: string
    }
    name: string
    sharing: Sharing
    skipTotal: boolean
    translations: Array<Translation>
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    user?: {
        id: string
    }
}

export namespace CategoryCombo {
    export enum dataDimensionType {
        DISAGGREGATION = 'DISAGGREGATION',
        ATTRIBUTE = 'ATTRIBUTE',
    }
}
// ...
export type DataElement = {
    access: Access
    aggregationLevels: Array<number>
    aggregationType: DataElement.aggregationType
    attributeValues: Array<AttributeValue>
    /**
     * A UID reference to a CategoryCombo
     * (Java name `org.hisp.dhis.category.CategoryCombo`)
     */
    categoryCombo?: {
        id: string
    }
    code: string
    /**
     * A UID reference to a OptionSet
     * (Java name `org.hisp.dhis.option.OptionSet`)
     */
    commentOptionSet?: {
        id: string
    }
    created: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    createdBy?: {
        id: string
    }
    dataElementGroups?: Array<{
        id: string
    }>
    dataSetElements: Array<DataSetElement>
    description: string
    dimensionItem: string
    displayDescription: string
    displayFormName: string
    displayName: string
    displayShortName: string
    domainType: DataElement.domainType
    favorite: boolean
    favorites: Array<string>
    fieldMask: string
    formName: string
    href: string
    id: string
    lastUpdated: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    lastUpdatedBy?: {
        id: string
    }
    /**
     * A UID reference to a LegendSet
     * (Java name `org.hisp.dhis.legend.LegendSet`)
     */
    legendSet?: {
        id: string
    }
    legendSets?: Array<{
        id: string
    }>
    name: string
    /**
     * A UID reference to a OptionSet
     * (Java name `org.hisp.dhis.option.OptionSet`)
     */
    optionSet?: {
        id: string
    }
    optionSetValue: boolean
    queryMods: QueryModifiers
    sharing: Sharing
    shortName: string
    style: ObjectStyle
    translations: Array<Translation>
    url: string
    /**
     * A UID reference to a User
     * (Java name `org.hisp.dhis.user.User`)
     */
    user?: {
        id: string
    }
    valueType: DataElement.valueType
    valueTypeOptions: FileTypeValueOptions
    zeroIsSignificant: boolean
}

export namespace DataElement {
    export enum aggregationType {
        SUM = 'SUM',
        AVERAGE = 'AVERAGE',
        AVERAGE_SUM_ORG_UNIT = 'AVERAGE_SUM_ORG_UNIT',
        LAST = 'LAST',
        LAST_AVERAGE_ORG_UNIT = 'LAST_AVERAGE_ORG_UNIT',
        LAST_LAST_ORG_UNIT = 'LAST_LAST_ORG_UNIT',
        LAST_IN_PERIOD = 'LAST_IN_PERIOD',
        LAST_IN_PERIOD_AVERAGE_ORG_UNIT = 'LAST_IN_PERIOD_AVERAGE_ORG_UNIT',
        FIRST = 'FIRST',
        FIRST_AVERAGE_ORG_UNIT = 'FIRST_AVERAGE_ORG_UNIT',
        FIRST_FIRST_ORG_UNIT = 'FIRST_FIRST_ORG_UNIT',
        COUNT = 'COUNT',
        STDDEV = 'STDDEV',
        VARIANCE = 'VARIANCE',
        MIN = 'MIN',
        MAX = 'MAX',
        MIN_SUM_ORG_UNIT = 'MIN_SUM_ORG_UNIT',
        MAX_SUM_ORG_UNIT = 'MAX_SUM_ORG_UNIT',
        NONE = 'NONE',
        CUSTOM = 'CUSTOM',
        DEFAULT = 'DEFAULT',
    }

    export enum domainType {
        AGGREGATE = 'AGGREGATE',
        TRACKER = 'TRACKER',
    }

    export enum valueType {
        TEXT = 'TEXT',
        LONG_TEXT = 'LONG_TEXT',
        MULTI_TEXT = 'MULTI_TEXT',
        LETTER = 'LETTER',
        PHONE_NUMBER = 'PHONE_NUMBER',
        EMAIL = 'EMAIL',
        BOOLEAN = 'BOOLEAN',
        TRUE_ONLY = 'TRUE_ONLY',
        DATE = 'DATE',
        DATETIME = 'DATETIME',
        TIME = 'TIME',
        NUMBER = 'NUMBER',
        UNIT_INTERVAL = 'UNIT_INTERVAL',
        PERCENTAGE = 'PERCENTAGE',
        INTEGER = 'INTEGER',
        INTEGER_POSITIVE = 'INTEGER_POSITIVE',
        INTEGER_NEGATIVE = 'INTEGER_NEGATIVE',
        INTEGER_ZERO_OR_POSITIVE = 'INTEGER_ZERO_OR_POSITIVE',
        TRACKER_ASSOCIATE = 'TRACKER_ASSOCIATE',
        USERNAME = 'USERNAME',
        COORDINATE = 'COORDINATE',
        ORGANISATION_UNIT = 'ORGANISATION_UNIT',
        REFERENCE = 'REFERENCE',
        AGE = 'AGE',
        URL = 'URL',
        FILE_RESOURCE = 'FILE_RESOURCE',
        IMAGE = 'IMAGE',
        GEOJSON = 'GEOJSON',
    }
}
// ...
    export enum style {
        COMFORTABLE = 'COMFORTABLE',
        NORMAL = 'NORMAL',
        COMPACT = 'COMPACT',
        NONE = 'NONE',
    }
// ...
export type DataSetElement = {
    /**
     * A UID reference to a CategoryCombo
     * (Java name `org.hisp.dhis.category.CategoryCombo`)
     */
    categoryCombo?: {
        id: string
    }
    /**
     * A UID reference to a DataElement
     * (Java name `org.hisp.dhis.dataelement.DataElement`)
     */
    dataElement?: {
        id: string
    }
    /**
     * A UID reference to a DataSet
     * (Java name `org.hisp.dhis.dataset.DataSet`)
     */
    dataSet?: {
        id: string
    }
}
// ...
export type FileTypeValueOptions = {
    allowedContentTypes: Array<string>
    maxFileSize: number
    version: number
}
// ...
export type ObjectStyle = {
    color: string
    icon: string
}
// ...
export type QueryModifiers = {
    aggregationType: QueryModifiers.aggregationType
    maxDate: string
    minDate: string
    periodOffset: number
    valueType: QueryModifiers.valueType
    yearToDate: boolean
}

export namespace QueryModifiers {
    export enum aggregationType {
        SUM = 'SUM',
        AVERAGE = 'AVERAGE',
        AVERAGE_SUM_ORG_UNIT = 'AVERAGE_SUM_ORG_UNIT',
        LAST = 'LAST',
        LAST_AVERAGE_ORG_UNIT = 'LAST_AVERAGE_ORG_UNIT',
        LAST_LAST_ORG_UNIT = 'LAST_LAST_ORG_UNIT',
        LAST_IN_PERIOD = 'LAST_IN_PERIOD',
        LAST_IN_PERIOD_AVERAGE_ORG_UNIT = 'LAST_IN_PERIOD_AVERAGE_ORG_UNIT',
        FIRST = 'FIRST',
        FIRST_AVERAGE_ORG_UNIT = 'FIRST_AVERAGE_ORG_UNIT',
        FIRST_FIRST_ORG_UNIT = 'FIRST_FIRST_ORG_UNIT',
        COUNT = 'COUNT',
        STDDEV = 'STDDEV',
        VARIANCE = 'VARIANCE',
        MIN = 'MIN',
        MAX = 'MAX',
        MIN_SUM_ORG_UNIT = 'MIN_SUM_ORG_UNIT',
        MAX_SUM_ORG_UNIT = 'MAX_SUM_ORG_UNIT',
        NONE = 'NONE',
        CUSTOM = 'CUSTOM',
        DEFAULT = 'DEFAULT',
    }

    export enum valueType {
        TEXT = 'TEXT',
        LONG_TEXT = 'LONG_TEXT',
        MULTI_TEXT = 'MULTI_TEXT',
        LETTER = 'LETTER',
        PHONE_NUMBER = 'PHONE_NUMBER',
        EMAIL = 'EMAIL',
        BOOLEAN = 'BOOLEAN',
        TRUE_ONLY = 'TRUE_ONLY',
        DATE = 'DATE',
        DATETIME = 'DATETIME',
        TIME = 'TIME',
        NUMBER = 'NUMBER',
        UNIT_INTERVAL = 'UNIT_INTERVAL',
        PERCENTAGE = 'PERCENTAGE',
        INTEGER = 'INTEGER',
        INTEGER_POSITIVE = 'INTEGER_POSITIVE',
        INTEGER_NEGATIVE = 'INTEGER_NEGATIVE',
        INTEGER_ZERO_OR_POSITIVE = 'INTEGER_ZERO_OR_POSITIVE',
        TRACKER_ASSOCIATE = 'TRACKER_ASSOCIATE',
        USERNAME = 'USERNAME',
        COORDINATE = 'COORDINATE',
        ORGANISATION_UNIT = 'ORGANISATION_UNIT',
        REFERENCE = 'REFERENCE',
        AGE = 'AGE',
        URL = 'URL',
        FILE_RESOURCE = 'FILE_RESOURCE',
        IMAGE = 'IMAGE',
        GEOJSON = 'GEOJSON',
    }
}
// ...
export type Sharing = {
    external: boolean
    owner: string
    public: string
    userGroups: Record<string, UserGroupAccess>
    users: Record<string, UserAccess>
}
// ...
export type Translation = {
    locale: string
    property: string
    value: string
}
// ...
export type UserAccess = {
    access: string
    displayName: string
    id: string
}
// ...
export type UserGroupAccess = {
    access: string
    displayName: string
    id: string
}